feat: add testing with pytest
This feature was inspired by the proper use of testing for CI/CD.

The structure was modified for accepting testing, and there were added 2 fixtures for http clients for testing:
- client: this is a normal http client that runs everything sync
- aclient: this is a special type of client for doing async calls

The tests should be properly marked with @pytest.mark.unit for unit tets, these should test no dependencies, but only internal functioning and are run when pushing a feature branch.

The tests that test integrations need to be marked with @pytest.mark.integration if it tests connection to other modules like the db or some external microservices.

The tests that test full workflows need to be marked with @pytest.mark.e2e